+++ lib/src/logging.dart
@@ -1,10 +1,8 @@
-// 📦 Package imports:
 import 'package:cli_util/cli_logging.dart';
-
-// 🌎 Project imports:
 import 'package:temptrail/src/config.dart';
 
 typedef LogCallback = void Function(String);
+typedef ColorWrapper = String Function(Object value);
 
 Logger logger = Logger.standard(ansi: Ansi(true));
 
@@ -20,3 +18,19 @@ extension LoggerUtils on Logger {
         : Logger.standard(ansi: ansi);
   }
 }
+
+void drawDivider() => log(subtle('-----------------------'));
+
+final noColor = _wrap(logger.ansi.noColor);
+final subtle = _wrap(logger.ansi.gray);
+final green = _wrap(logger.ansi.green);
+final red = _wrap(logger.ansi.red);
+final blue = _wrap(logger.ansi.blue);
+final magenta = _wrap(logger.ansi.magenta);
+final yellow = _wrap(logger.ansi.yellow);
+
+final none = _wrap(logger.ansi.none);
+final bold = _wrap(logger.ansi.bold);
+
+ColorWrapper _wrap(String ansiModifier) =>
+    (value) => '$ansiModifier$value${logger.ansi.none}';