+++ lib/src/utils.dart
@@ -1,7 +1,5 @@
-// 🎯 Dart imports:
 import 'dart:async';
 
-// 📦 Package imports:
 import 'package:meta/meta.dart';
 
 // Typedefs