+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0
+
+- Fixed bug where I forgot to implement "exit on error" behavior properly.
+- Added callback for exiting program that sets AnyBar to black.
+- Improved ANSI coloring.
+
 ## 1.1.0
 
 - First working version.